Added documentation for amazon_s3 adapter
I've added what I understand based on getting this working today. Not sure about all the options. 

I also added a note about bucket names with dots in them - this had me stuck for a while until I found mention of it in passing in an AWS ticket https://forums.aws.amazon.com/thread.jspa?threadID=122890#jive-message-445833

If it's not a generally repeatable problem feel free to delete that line, but it certainly caused me problems!